api/pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "doctr-api"
version = "1.0.1a0"
description = "Backend template for your OCR API with docTR"
authors = ["Mindee <contact@mindee.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-doctr = {git = "https://github.com/mindee/doctr.git", branch = "main" }
# Fastapi: minimum version required to avoid pydantic error
# cf. https://github.com/tiangolo/fastapi/issues/4168
fastapi = ">=0.73.0"
uvicorn = ">=0.11.1"
python-multipart = ">=0.0.5"
# Document format support
Pillow = "^10.0.0"
python-docx = "^1.1.0"
striprtf = "^0.0.26"
odfpy = "^1.4.1"
# AI extraction support (optional)
openai = {version = "^1.0.0", optional = true}
anthropic = {version = "^0.25.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = ">=5.3.2"
pytest-asyncio = ">=0.14.0"
httpx = ">=0.23.0"
requests = ">=2.20.0"
